+++ application/controllers/Stock.php
@@ -1,7 +1,7 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-class Welcome extends CI_Controller {
+class Stock extends CI_Controller {
 
 	/**
 	 * Index Page for this controller.
@@ -19,17 +19,7 @@ class Welcome extends CI_Controller {
 	 * @see https://codeigniter.com/user_guide/general/urls.html
 	 */
 	public function index()
-	{
-		$this->load->view('welcome_message');
-        /*
-        $this->load->database();
-        $query = $this->db->query('SELECT Player FROM players');
-
-        foreach ($query->result() as $row) {
-            echo $row->Player . "";
-        }
-        */
-        
-        
+	{   
+        $this->load->view("TempStock");
 	}
 }