+++ changelog.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented to this file.
 Change Log history will be recorded in descending order
 
 
+## 0.0.2
+** Changed Files **
+- Added Controller, View for Portfolio and Stock
+- Added Portfolio Model
+- Added Routes for Portfolio
+
 ## 0.0.1
 ** Added Files **
 - Added CodeIgniter Library, initial project established